Options to disable autoindex and dirzip features
Each allowed URL prefix can now take a list of options, separated by a
comma. e.g: --allow /url=/path/on/disk:noautoindex,nodirzip

nodirzip disables the main feature of this program, that is, when
directory_name.zip is requested a zip file with the content of the
directory  is generated on the fly.
With nodirzip, this feature is disabled.

noautoindex disables the directory listing when a directory is requested.

Both can be combined for a plain http file server with no index capability.